Add a "one-or-more" JSON reader: `reader/json+`
The `add` endpoint doesn't return a single JSON message with all the
CIDs but multiple JSON messages, one for each added file or directory.

Change the default reader for the `add` endpoint to `reader/json+`.

The `quiet` and `quieter` flags seem to do nothing on the `add`
endpoint; but the `silent` flag seems to be equivalent to the `ipfs`
command's `--silent`/`--quieter` flags.